Created an Environment Tree Class to store EnvNodes
Created a List that stores all the EnvNodes instead of tracking the current env only. Had to reformat the CSE class due to this reason. Environment nodes are not deleted but the access to them is lost as the current environment number gets updated. They can only be accessed if there are listed as a parent env only. Managed to fix CSE Rule5. But there are errors when handling recursion based programs